Confirm unsaved changes before logout

+++ src/components/navbar/actions.vue
@@ -40,6 +40,7 @@ except according to the terms contained in the LICENSE file.
 <script>
 import request from '../../mixins/request';
 import { apiPaths } from '../../util/request';
+import { confirmUnsavedChanges } from '../../util/router';
 import { noop } from '../../util/util';
 import { requestData } from '../../store/modules/request';
 
@@ -57,6 +58,7 @@ export default {
   computed: requestData(['currentUser', 'session']),
   methods: {
     logOut() {
+      if (!confirmUnsavedChanges(this.$store)) return;
       this.request({
         method: 'DELETE',
         url: apiPaths.session(this.session.token)
